Reorganize widget modules
