+++ naturtag/controllers/__init__.py
@@ -1,7 +1,7 @@
 # flake8: noqa: F401
 # isort: skip_file
-from naturtag.controllers.image_gallery import ImageGallery, ImageWindow
-from naturtag.controllers.taxon_view import TaxonInfoCard, TaxonInfoSection, TaxonList, TaxonomySection
+from naturtag.controllers.image_gallery import ImageGallery
 from naturtag.controllers.image_controller import ImageController
 from naturtag.controllers.taxon_search import TaxonSearch
+from naturtag.controllers.taxon_view import TaxonInfoSection, TaxonomySection
 from naturtag.controllers.taxon_controller import TaxonController
