+++ naturtag/controllers/image_gallery.py
@@ -7,8 +7,8 @@ from typing import Callable, Iterable
 from urllib.parse import unquote, urlparse
 
 from PySide6.QtCore import Qt, QUrl, Signal, Slot
-from PySide6.QtGui import QAction, QDesktopServices, QDropEvent, QKeySequence, QPixmap, QShortcut
-from PySide6.QtWidgets import QApplication, QFileDialog, QLabel, QMenu, QScrollArea, QWidget
+from PySide6.QtGui import QAction, QDesktopServices, QDropEvent, QPixmap
+from PySide6.QtWidgets import QApplication, QFileDialog, QLabel, QMenu, QScrollArea
 
 from naturtag.app.style import fa_icon
 from naturtag.constants import IMAGE_FILETYPES, THUMBNAIL_SIZE_DEFAULT
@@ -19,7 +19,7 @@ from naturtag.widgets import (
     FlowLayout,
     HorizontalLayout,
     IconLabel,
-    PixmapLabel,
+    ImageWindow,
     StylableWidget,
     VerticalLayout,
 )
@@ -115,61 +115,6 @@ class ImageGallery(StylableWidget):
     @Slot(str)
     def select_image(self, file_path: str):
         self.image_window.select_image(file_path, list(self.images.keys()))
-
-
-# TODO: Small overlay with photo info
-class ImageWindow(QWidget):
-    """Display a single full-size image at a time as a separate window
-
-    Keyboard shortcuts: Escape to close window, Left and Right to cycle through images
-    """
-
-    def __init__(self):
-        super().__init__()
-        self.image_paths: list[str] = []
-        self.selected_path = None
-
-        self.image = PixmapLabel()
-        self.image.setFixedSize(QApplication.primaryScreen().availableSize())
-        self.image.setAlignment(Qt.AlignCenter)
-        self.image_layout = VerticalLayout(self)
-        self.image_layout.addWidget(self.image)
-        self.image_layout.setContentsMargins(0, 0, 0, 0)
-
-        # Keyboard shortcuts
-        shortcut = QShortcut(QKeySequence(Qt.Key_Escape), self)
-        shortcut.activated.connect(self.close)
-        shortcut = QShortcut(QKeySequence(Qt.Key_Right), self)
-        shortcut.activated.connect(self.select_next_image)
-        shortcut = QShortcut(QKeySequence(Qt.Key_Left), self)
-        shortcut.activated.connect(self.select_prev_image)
-
-    def select_image(self, selected_path: str, image_paths: list[str]):
-        """Open window to a selected image, and save other available image paths for navigation"""
-        self.selected_path = selected_path
-        self.image_paths = image_paths
-        self.set_pixmap(self.selected_path)
-        self.showFullScreen()
-
-    def select_image_idx(self, idx: int):
-        """Select an image by index, with wraparound"""
-        if idx < 0:
-            idx = len(self.image_paths) - 1
-        elif idx >= len(self.image_paths):
-            idx = 0
-
-        logger.debug(f'Selecting image {idx}: {self.selected_path}')
-        self.selected_path = self.image_paths[idx]
-        self.set_pixmap(self.selected_path)
-
-    def select_next_image(self):
-        self.select_image_idx(self.image_paths.index(self.selected_path) + 1)
-
-    def select_prev_image(self):
-        self.select_image_idx(self.image_paths.index(self.selected_path) - 1)
-
-    def set_pixmap(self, path: str):
-        self.image.setPixmap(QPixmap(path))
 
 
 class LocalThumbnail(StylableWidget):
