+++ naturtag/widgets/__init__.py
@@ -11,6 +11,7 @@ from naturtag.widgets.layouts import (
     VerticalLayout,
 )
 from naturtag.widgets.autocomplete import TaxonAutocomplete
-from naturtag.widgets.images import IconLabel, PixmapLabel, fetch_image
+from naturtag.widgets.images import IconLabel, ImageWindow, PixmapLabel, fetch_image
 from naturtag.widgets.logger import QtRichHandler, init_handler
+from naturtag.widgets.taxon_images import TaxonImageWindow, TaxonInfoCard, TaxonList, TaxonPixmapLabel
 from naturtag.widgets.toggle_switch import ToggleSwitch
