+++ naturtag/widgets/images.py
@@ -1,14 +1,16 @@
+"""Generic image widgets"""
 from logging import getLogger
 from pathlib import Path
 from typing import Union
 
 from pyinaturalist import Photo, Taxon
 from PySide6.QtCore import QSize, Qt
-from PySide6.QtGui import QPixmap
-from PySide6.QtWidgets import QLabel, QWidget
+from PySide6.QtGui import QKeySequence, QPixmap, QShortcut
+from PySide6.QtWidgets import QApplication, QLabel, QWidget
 
 from naturtag.app.style import fa_icon
 from naturtag.client import IMG_SESSION
+from naturtag.widgets import VerticalLayout
 
 logger = getLogger(__name__)
 
@@ -78,6 +80,61 @@ class PixmapLabel(QLabel):
     def resizeEvent(self, _):
         if self._pixmap:
             super().setPixmap(self.scaledPixmap())
+
+
+# TODO: Small overlay with photo info
+class ImageWindow(QWidget):
+    """Display a single full-size image at a time as a separate window
+
+    Keyboard shortcuts: Escape to close window, Left and Right to cycle through images
+    """
+
+    def __init__(self):
+        super().__init__()
+        self.image_paths: list[str] = []
+        self.selected_path = None
+
+        self.image = PixmapLabel()
+        self.image.setFixedSize(QApplication.primaryScreen().availableSize())
+        self.image.setAlignment(Qt.AlignCenter)
+        self.image_layout = VerticalLayout(self)
+        self.image_layout.addWidget(self.image)
+        self.image_layout.setContentsMargins(0, 0, 0, 0)
+
+        # Keyboard shortcuts
+        shortcut = QShortcut(QKeySequence(Qt.Key_Escape), self)
+        shortcut.activated.connect(self.close)
+        shortcut = QShortcut(QKeySequence(Qt.Key_Right), self)
+        shortcut.activated.connect(self.select_next_image)
+        shortcut = QShortcut(QKeySequence(Qt.Key_Left), self)
+        shortcut.activated.connect(self.select_prev_image)
+
+    def select_image(self, selected_path: str, image_paths: list[str]):
+        """Open window to a selected image, and save other available image paths for navigation"""
+        self.selected_path = selected_path
+        self.image_paths = image_paths
+        self.set_pixmap(self.selected_path)
+        self.showFullScreen()
+
+    def select_image_idx(self, idx: int):
+        """Select an image by index, with wraparound"""
+        if idx < 0:
+            idx = len(self.image_paths) - 1
+        elif idx >= len(self.image_paths):
+            idx = 0
+
+        logger.debug(f'Selecting image {idx}: {self.selected_path}')
+        self.selected_path = self.image_paths[idx]
+        self.set_pixmap(self.selected_path)
+
+    def select_next_image(self):
+        self.select_image_idx(self.image_paths.index(self.selected_path) + 1)
+
+    def select_prev_image(self):
+        self.select_image_idx(self.image_paths.index(self.selected_path) - 1)
+
+    def set_pixmap(self, path: str):
+        self.image.setPixmap(QPixmap(path))
 
 
 def fetch_image(photo: Photo, size: str = None) -> QPixmap:
